+++ build.properties
@@ -1,4 +1,4 @@
-#Sat, 25 Apr 2020 13:58:12 +0300
+#Sat, 25 Apr 2020 15:36:45 +0300
 project.name=lab
 
 lib.dir=lib
@@ -6,9 +6,9 @@ lib.junit=lib/junit-4.12.jar
 lib.hamcrest=lib/hamcrest-core-1.3.jar
 
 src.dir=src
-src.dir.doc=src
-src.res=META-INF
-src.test=src/tests
+src.dir.doc=src/main/java
+src.res=src/main/resources/META-INF
+src.test=src/test/java
 web.content=web
 
 build.dir=ant/build
