+++ src/main/java/beans/MatchingManager.java
@@ -1,4 +1,4 @@
-package beans;
+package main.java.beans;
 
 /**
  * Method for testing javadoc
