+++ src/main/java/database/JDBCManager.java
@@ -1,4 +1,4 @@
-package database;
+package main.java.database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
