+++ src/main/java/validator/XValidator.java
@@ -1,4 +1,4 @@
-package validator;
+package main.java.validator;
 
 import org.primefaces.context.RequestContext;
 
@@ -8,26 +8,28 @@ import javax.faces.context.FacesContext;
 import javax.faces.validator.FacesValidator;
 import javax.faces.validator.Validator;
 import javax.faces.validator.ValidatorException;
-import java.util.logging.Logger;
 
-@FacesValidator("UValidator")
-public class YValidator implements Validator {
-    private static final double minY = -5;
-    private static final double maxY = 5;
+@FacesValidator("XValidator")
+public class XValidator implements Validator {
+    private static final double minX = -2;
+    private static final double maxX = 2;
 
     public void validate(FacesContext facesContext, UIComponent uiComponent, Object o) throws ValidatorException {
         try {
-            double y = Double.parseDouble(o.toString().replace(',', '.'));
-            if (!(y >= minY && y <= maxY))
+
+            double x = Double.parseDouble(o.toString().replace(',', '.'));
+            if (!(x >= minX && x <= maxX))
                 throw new IllegalArgumentException();
+
         } catch (Exception e) {
             FacesMessage msg =
-                    new FacesMessage("Y validation failed.",
-                            "\u041D\u0435\u0432\u0435\u0440\u043D\u0430\u044F \20\u043A\u043E\u043E\u0440\u0434\u0438\u043D\u0430\u0442\u0430\20 Y."
+                    new FacesMessage("X validation failed.",
+                            "\u041D\u0435\u0432\u0435\u0440\u043D\u0430\u044F \20\u043A\u043E\u043E\u0440\u0434\u0438\u043D\u0430\u0442\u0430\20 X."
                     );
             msg.setSeverity(FacesMessage.SEVERITY_ERROR);
             RequestContext.getCurrentInstance().showMessageInDialog(msg);
             throw new ValidatorException(msg);
         }
     }
+
 }
